fix: 불안 온도 리포트 소수점 응답 처리

+++ app/src/main/java/com/gominitta/android/data/remote/dto/AnxietyGapResponse.kt
@@ -5,9 +5,9 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class AnxietyGapResponse(
     val period: String,
-    val beforeScore: Long,
-    val afterScore: Long,
-    val gap: Long,
+    val beforeScore: Double,
+    val afterScore: Double,
+    val gap: Double,
     val sampleCount: Long,
     val feedback: String,
 )
